Moved calculatingWinner to separate file and added tests for card conversion
